% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{all_in}
\alias{all_in}
\title{Fonction all_in. Applique rmeaps dans sa version "tous les actifs partent en même temps".}
\usage{
all_in(
  MeapsData,
  attraction = "constant",
  parametres = 0,
  nthreads = 0L,
  verbose = TRUE
)
}
\arguments{
\item{MeapsData}{Un jeu de données MeapsData décrivant la zone d'étude.}

\item{attraction}{Choix du mode d'attraction. Par défaut, "constant". Les autres modes possibles sont : "marche", "marche_liss", "decay" et "logistique". Chacune applique une pénalité selon la distance (au-delà de la simple logique des rangs, qui correspond au mode "constant").}

\item{parametres}{Un vecteur avec les valeurs initiales des paramètres. Dépend du mode d'attraction retenu.}

\item{nthreads}{Le nombre de threads pour mener les calculs parallélisés. Par defaut, le nombre maximal.}

\item{verbose}{TRUE par defaut.}
}
\description{
Fonction all_in. Applique rmeaps dans sa version "tous les actifs partent en même temps".
}
