% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{anderson_darling}
\alias{anderson_darling}
\title{Test de Anderson–Darling
on suppose que les flux sont classés}
\usage{
anderson_darling(est, obs, w = 1)
}
\arguments{
\item{est}{distribution empirique estimée}

\item{obs}{distribution empirique observée}
}
\value{
un double, le AD
}
\description{
Test de Anderson–Darling
on suppose que les flux sont classés
}
