% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chances_absorption}
\alias{chances_absorption}
\title{La fonction renvoie les chances d'absorption issue de meaps.}
\usage{
chances_absorption(rkdist, emplois, modds, f)
}
\arguments{
\item{rkdist}{La matrice des rangs dans lequel les colonnes j sont passées en revue pour chacune des lignes i.}

\item{emplois}{Le vecteur des emplois disponibles sur chacun des sites j (= marge des colonnes).}

\item{modds}{La matrice des odds modifiant la chance d'absorption de chacun des sites j pour des résidents en i.}

\item{f}{Le vecteur de la probabilité de fuite des actifs hors de la zone d'étude.}
}
\value{
renvoie le vecteur des chances d'absorption.
}
\description{
La fonction renvoie les chances d'absorption issue de meaps.
}
