% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{communaliser}
\alias{communaliser}
\title{La fonction communaliser effectue un regroupement de la matrice des flux (résultant de meaps) selon des groupes origines et destinations.}
\usage{
communaliser(flux, group_orig, group_dest)
}
\arguments{
\item{flux}{La matrice des flux de fromid vers toid.}

\item{group_orig}{Un vecteur d'integers (par ex. code commune) donnant le code du groupe de départ.}

\item{group_dest}{Un vecteur d'integers (par ex. code commune) donnant le code du groupe d'arrivée.}
}
\value{
Une matrice contenant les flux agrégés selon les groupes de départ et d'arrivée.
}
\description{
La fonction communaliser effectue un regroupement de la matrice des flux (résultant de meaps) selon des groupes origines et destinations.
}
