% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cramer_vonmises}
\alias{cramer_vonmises}
\title{Test de Cramer–von Mises
on suppose que les flux sont classés}
\usage{
cramer_vonmises(est, obs, w = 1)
}
\arguments{
\item{est}{distribution empirique estimée}

\item{obs}{distribution empirique observée}
}
\value{
un double, le AD
}
\description{
Test de Cramer–von Mises
on suppose que les flux sont classés
}
