% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deborder}
\alias{deborder}
\title{La fonction déborder remplit une série de conteneurs dans l'ordre de présentation. Il passe au conteneur suivant si le précédent est rempli.
Celle-ci est exportée et donne le reste non distribué.}
\usage{
deborder(conteneurs, quantite)
}
\arguments{
\item{conteneurs}{Un vecteur donnant la capacité de chacun des conteneurs dans l'ordre voulu de remplissage.}

\item{quantité}{La quantité totale à verser dans les conteneurs.}
}
\value{
Une liste avec : part = la répartition finale de la quantité dans les conteneurs et reste = la quantité qui n'a pas pu être versée si tout est rempli.
}
\description{
La fonction déborder remplit une série de conteneurs dans l'ordre de présentation. Il passe au conteneur suivant si le précédent est rempli.
Celle-ci est exportée et donne le reste non distribué.
}
