% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distribuer}
\alias{distribuer}
\title{La fonction distribuer remplit un ensemble de conteneurs. 
Celle-ci est exportée et donne également le reste non distribué.}
\usage{
distribuer(conteneurs, proportion, quantite)
}
\arguments{
\item{conteneurs}{Un vecteur donnant la capacité de chacun des conteneurs dans l'ordre voulu de remplissage.}

\item{proportion}{Un vecteur définissant les proportions des débits vers les conteneurs.}

\item{quantité}{La quantité totale à verser dans les conteneurs.}
}
\value{
Une liste avec : part = la répartition finale de la quantité dans les conteneurs et reste = la quantité qui n'a pas pu être versée si tout est rempli.
}
\description{
La fonction distribuer remplit un ensemble de conteneurs. 
Celle-ci est exportée et donne également le reste non distribué.
}
