% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.matrix2dgR}
\alias{.matrix2dgR}
\title{Fonction de retraitement de la distance si elle est sous forme d'une matrice classique.
Les NA deviennent sparses. Les zéros deviennent simplement des petites valeurs.
Le résultat est une dgRMatrix.}
\usage{
.matrix2dgR(x, na2zero = TRUE)
}
\description{
Fonction de retraitement de la distance si elle est sous forme d'une matrice classique.
Les NA deviennent sparses. Les zéros deviennent simplement des petites valeurs.
Le résultat est une dgRMatrix.
}
