% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.meaps_continu}
\alias{.meaps_continu}
\title{La fonction meaps en mode continu sur plusieurs shufs avec en entrée une Row Sparse Matrix destructurée selon ses éléments.}
\usage{
.meaps_continu(
  j_dist,
  p_dist,
  x_dist,
  emplois,
  actifs,
  f,
  shuf,
  param,
  j_odds,
  p_odds,
  x_odds,
  attraction = "constant",
  nthreads = 0L,
  progress = TRUE,
  normalisation = FALSE,
  fuite_min = 0.001
)
}
\arguments{
\item{j_dist}{Le vecteur des indices des colonnes non vides.}

\item{p_dist}{Le vecteur du nombres de valeurs non nulles sur chacune des lignes.}

\item{x_dist}{Le vecteur des valeurs dans l'ordre de j_dist.}

\item{emplois}{Le vecteur des emplois disponibles sur chacun des sites j (= marge des colonnes).}

\item{actifs}{Le vecteur des actifs partant de chacune des lignes visées par shuf. Le vecteur doit faire la même longueur que shuf.}

\item{f}{Le vecteur de la probabilité de fuite des actifs hors de la zone d'étude.}

\item{shuf}{Le vecteur de priorité des actifs pour choisir leur site d'arrivée. Il est possible de segmenter les départs d'une ligne i 
en répétant cette ligne à plusieurs endroits du shuf et en répartissant les poids au sein du vecteurs actifs.}

\item{param}{est un vecteur avec dans l'ordre les valeurs des paramètres.}

\item{j_odds, }{p_odds et x_odds sont les vecteurs de la Row Sparse Matrix lorsque attraction = "odds".}

\item{attraction}{Choix de la fonction d'attraction des différents sites, appliquée à l'accessibilité. 
Par défaut, "constant" où aucun site n'a plus d'attrait qu'un autre. 
"marche" où l'attrait vaut 1 jusqu'à une certaine distance (param 1) puis moins (param 2). f(x) = 1 si x < p1, = p2 si x > p1.
"logistique" où l'attrait décroît selon une fonction logistique avec une distance de bascule (param 1), une vitesse de bascule (param 2) 
et un seuil (param p). Si h(x) = exp( (x-p1)/p2), f(x) = p3 + h(x) / (1 + h(x)).
"odds" où chaque flux (from, to) se voit attribuer un odds. Dans ce cas, on entre un Row Sparse Matrix des log(odds) selon ses éléments.}

\item{nthreads}{Nombre de threads pour OpenMP. Default : 0 = choix auto.}

\item{progress}{Ajoute une barre de progression. Default : true.}

\item{normalisation}{Calage des emplois disponibles sur le nombre d'actifs travaillant sur la zone. Default : false.}

\item{fuite_min}{Seuil minimal pour la fuite d'un actif. Doit être supérieur à 0. Défault = 1e-3.}
}
\value{
renvoie un vecteur des estimations des flux de i vers j.
}
\description{
La fonction meaps en mode continu sur plusieurs shufs avec en entrée une Row Sparse Matrix destructurée selon ses éléments.
}
