% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R, R/utils.R
\name{.petite_distance}
\alias{.petite_distance}
\title{Fonction de choix d'une petite distance pour remplacer les zéros éventuels.
Normalement il ne devrait pas y avoir de zéros. 
Au sein d'un même carreau de dim a, la distance entre deux points est d'environ a/2.
La distance retenue importe peu tant que les rangs sont respectés.}
\usage{
.petite_distance(d, facteur = 10)

.petite_distance(d, facteur = 10)
}
\description{
Fonction de choix d'une petite distance pour remplacer les zéros éventuels.
Normalement il ne devrait pas y avoir de zéros. 
Au sein d'un même carreau de dim a, la distance entre deux points est d'environ a/2.
La distance retenue importe peu tant que les rangs sont respectés.

Fonction de choix d'une petite distance pour remplacer les zéros éventuels.
Normalement il ne devrait pas y avoir de zéros. 
Au sein d'un même carreau de dim a, la distance entre deux points est d'environ a/2.
La distance retenue importe peu tant que les rangs sont respectés.
}
