% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{.transfom_matrix}
\alias{.transfom_matrix}
\title{Fonction de retraitement de la distance si elle est sous forme d'une matrice classique.
Les NA doivent devenir sparses. Les zéros deviennent simplement des petites valeurs.}
\usage{
.transfom_matrix(dist)
}
\description{
Fonction de retraitement de la distance si elle est sous forme d'une matrice classique.
Les NA doivent devenir sparses. Les zéros deviennent simplement des petites valeurs.
}
