% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{.transform_triplet}
\alias{.transform_triplet}
\title{Fonction de retraitement de la distance si elle est sous forme de triplet dans une liste.
C'est en général la résultante de la fonction Matrix::mat2triplet.
On suppose que max(j) donne le nombre de colonnes, ce qui est attendu dans les analyses meaps.
Ici encore les valeurs x nulles deviennent des petites distances.
Liste de trois vecteurs : i, j et x.}
\usage{
.transform_triplet(dist)
}
\description{
Fonction de retraitement de la distance si elle est sous forme de triplet dans une liste.
C'est en général la résultante de la fonction Matrix::mat2triplet.
On suppose que max(j) donne le nombre de colonnes, ce qui est attendu dans les analyses meaps.
Ici encore les valeurs x nulles deviennent des petites distances.
Liste de trois vecteurs : i, j et x.
}
