% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiette.R
\name{emiette}
\alias{emiette}
\title{Emiette les actifs et produit un jeu d'arrangements (shuffles) des actifs}
\usage{
emiette(
  les_actifs,
  nshuf = 256,
  seuil = 40,
  var = "actifs",
  weighted = TRUE,
  seed = NULL
)
}
\arguments{
\item{les_actifs}{le tibble ou le vecteur des actifs (nommé par idINS)}

\item{nshuf}{le nombre de tirages}

\item{seuil}{le seuil d'emmiettage}

\item{var}{si les actifs sont un tibble, le nom de la variable 'actifs' ('actifs' par défaut)}

\item{weighted}{si TRUE, les actifs sont pondérés par leur fréquence (TRUE par défaut)}
}
\value{
une matrice d'arrangement (autant de lignes que nshuf, autant de colonnes que d'actifs, dont les colonnes sont l'idINS)
}
\description{
Emiette les actifs et produit un jeu d'arrangements (shuffles) des actifs
}
