% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genere_data.r
\name{genere_data}
\alias{genere_data}
\title{Genere des données pour les tests}
\usage{
genere_data(
  n = 3,
  k = 3,
  nshuf = 64,
  nb_actifs = n * n * 10,
  nb_emplois = nb_actifs * (1 - fuite),
  fuite = 0.1,
  densite = "uniforme",
  ties = "random"
)
}
\arguments{
\item{n}{nombre de points de résidents (sur une grille 0:1x0:1) défaut : 2}

\item{k}{nombre de points d'emplois (sur une grille 0:1x0:1) défaut : 2}

\item{nshuf}{nombre de tirages défaut : 64}

\item{nb_actifs}{nombre d'actifs défaut : 10xn^2}

\item{nb_emplois}{nombre d'emplois défaut : 9xk^2}

\item{fuite}{fuite défaut : 10%}

\item{densite}{forme de la répartition des actifs et des emplois. Si "uniforme" alors c'est uniforme, "radial" c'ets en 1/r^2 ou r est la distance au centre}
}
\value{
une liste pour de données meaps (rkdist les rangs, dist les distances, actifs les actifs, emplois les emplois, shuf un shuf)
}
\description{
Genere des données pour les tests
}
\examples{
genere_data(densite="radiale")
}
