% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\name{initialize,MeapsData-method}
\alias{initialize,MeapsData-method}
\title{Méthode d'initialisation pour MeapsData.}
\usage{
\S4method{initialize}{MeapsData}(
  .Object,
  triplet,
  actifs,
  emplois,
  fuites,
  nshuf = NULL,
  seuil = 40,
  seed = NULL,
  quiet = FALSE
)
}
\description{
Méthode d'initialisation pour MeapsData.
}
