% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\name{initialize,MeapsDataGroup-method}
\alias{initialize,MeapsDataGroup-method}
\title{Méthode d'initialisation pour MeapsDataGroup.}
\usage{
\S4method{initialize}{MeapsDataGroup}(
  .Object,
  triplet,
  actifs,
  emplois,
  fuites,
  shuf,
  froms,
  tos,
  group_from,
  group_to,
  cible,
  jr_dist,
  p_dist,
  quiet = FALSE
)
}
\description{
Méthode d'initialisation pour MeapsDataGroup.
}
