% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\name{initialize,MeapsDataMultiChamps-method}
\alias{initialize,MeapsDataMultiChamps-method}
\title{Méthode d'initialisation pour MeapsDataMultiChamps.
df_group est une data.frame avec fromidINS, toidINS classé comme le triplet, puis des variables de regroupements.}
\usage{
\S4method{initialize}{MeapsDataMultiChamps}(
  .Object,
  triplet,
  actifs,
  emplois,
  fuites,
  shuf,
  froms,
  tos,
  df_groups,
  cibles,
  jr_dist,
  p_dist
)
}
\description{
Méthode d'initialisation pour MeapsDataMultiChamps.
df_group est une data.frame avec fromidINS, toidINS classé comme le triplet, puis des variables de regroupements.
}
