% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_triplet}
\alias{is_triplet}
\title{utils}
\usage{
is_triplet(obj)
}
\description{
Fonction de test de la validité d'un triplet sous forme de liste ou de data.frame.
Nom de variables admises : i ou fromidINS, j ou toidINS.
Le nom de la troisième variable est libre.
}
