% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{khi2}
\alias{khi2}
\title{Un bête khi2}
\usage{
khi2(estime, observe)
}
\arguments{
\item{estime}{distribution empirique estimée}

\item{observe}{distribution empirique observée}
}
\value{
khi2
}
\description{
Un bête khi2
}
