% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_fluxes.r
\name{kl_fluxes}
\alias{kl_fluxes}
\title{Calcule la kl-divergence entre des flux "aux carreaux" et une matrice de flux "communaux" de référence.}
\usage{
kl_fluxes(
  flux,
  flux_ref,
  group_orig = NULL,
  group_dest = NULL,
  seuil_collapse = 0.1
)
}
\arguments{
\item{flux}{Matrice de flux, typiquement issue de meaps.}

\item{flux_ref}{Matrice de flux regroupés par catégories (typiquement des communes), servant de référence.}

\item{group_orig}{Vecteur désignant la catégorie des "carreaux" de départ. Si null, les rownames de flux sont utilisées.}

\item{group_dest}{Vecteur désignant la catégorie des "carreaux" d'arrivée. Si null, les colnames de flux sont utilisées.}
}
\value{
Retourne la valeur de Kullback-Leibler.
}
\description{
Calcule la kl-divergence entre des flux "aux carreaux" et une matrice de flux "communaux" de référence.
}
