% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{kolmogorov_smirnov}
\alias{kolmogorov_smirnov}
\title{Test de Kolmogorov-Smirnov
on suppose que les flux sont classés}
\usage{
kolmogorov_smirnov(est, obs)
}
\arguments{
\item{est}{distribution empirique estimée}

\item{obs}{distribution empirique observée}
}
\value{
un double, le KS
}
\description{
Test de Kolmogorov-Smirnov
on suppose que les flux sont classés
}
