% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ks}
\alias{ks}
\title{Test de Kolmogorov-Smirnov
on suppose que les flux sont classés}
\usage{
ks(est, obs)
}
\arguments{
\item{est}{distribution empirique estimée}

\item{obs}{distribution empirique observée}
}
\value{
ks
}
\description{
Test de Kolmogorov-Smirnov
on suppose que les flux sont classés
}
