% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listij2dgR}
\alias{listij2dgR}
\title{Fonction de retraitement de la distance si elle est sous forme de triplet dans une liste.
C'est en général la résultante de la fonction Matrix::mat2triplet.
On suppose que max(j) donne le nombre de colonnes, ce qui est attendu dans les analyses meaps.
On ne teste pas la validité du triplet.
Ici encore les valeurs x nulles deviennent des petites distances.
Liste de trois vecteurs : i, j et x.}
\usage{
listij2dgR(obj)
}
\description{
Fonction de retraitement de la distance si elle est sous forme de triplet dans une liste.
C'est en général la résultante de la fonction Matrix::mat2triplet.
On suppose que max(j) donne le nombre de colonnes, ce qui est attendu dans les analyses meaps.
On ne teste pas la validité du triplet.
Ici encore les valeurs x nulles deviennent des petites distances.
Liste de trois vecteurs : i, j et x.
}
