% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meaps_alt}
\alias{meaps_alt}
\title{La fonction meaps sur plusieurs shufs
version optimisée}
\usage{
meaps_alt(
  rkdist,
  emplois,
  actifs,
  modds,
  f,
  shuf,
  mode = "continu",
  oddssubjectifs = NULL,
  nthreads = 0L,
  progress = TRUE,
  normalisation = FALSE,
  fuite_min = 0.001,
  seuil_newton = 1e-06
)
}
\arguments{
\item{rkdist}{La matrice des rangs dans lequel les colonnes j sont passées en revue pour chacune des lignes i.}

\item{emplois}{Le vecteur des emplois disponibles sur chacun des sites j (= marge des colonnes).}

\item{actifs}{Le vecteur des actifs partant de chacune des lignes visées par shuf. Le vecteur doit faire la même longueur que shuf.}

\item{modds}{La matrice des odds modifiant la chance d'absorption de chacun des sites j pour des résidents en i.}

\item{f}{Le vecteur de la probabilité de fuite des actifs hors de la zone d'étude.}

\item{shuf}{Le vecteur de priorité des actifs pour choisir leur site d'arrivée. Il est possible de segmenter les départs d'une ligne i en répétant cette ligne à plusieurs endroits du shuf et en répartissant les poids au sein du vecteurs actifs.}

\item{mode}{Choix du rôle de l'emploi disponible au cours du processus. Default : continu. Autre choix : discret, subjectif_c ou _d...}

\item{nthreads}{Nombre de threads pour OpenMP. Default : 0 = choix auto.}

\item{progress}{Ajoute une barre de progression. Default : true.}

\item{normalisation}{Calage des emplois disponibles sur le nombre d'actifs travaillant sur la zone. Default : false.}

\item{fuite_min}{Seuil minimal pour la fuite d'un actif. Doit être supérieur à 0. Défault = 1e-3.}

\item{odds_subjectifs}{Attractivité des sites proches des actifs, pour le mode defini. default : null.}
}
\value{
renvoie une matrice avec les estimations du nombre de trajets de i vers j.
}
\description{
La fonction meaps sur plusieurs shufs
version optimisée
}
