% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_continu.R
\name{meaps_continu}
\alias{meaps_continu}
\title{La fonction meaps sur plusieurs shufs selon le mode de calcul continu.}
\usage{
meaps_continu(
  dist,
  emplois,
  actifs,
  f,
  shuf,
  attraction = "constant",
  param = numeric(),
  modds = NULL,
  nthreads = 0,
  progress = TRUE,
  normalisation = FALSE,
  fuite_min = 0.001,
  seuil_newton = 1e-06,
  quiet = FALSE
)
}
\arguments{
\item{dist}{un triplet sous forme de data.frame avec des colonnes fromidINS et toidINS.}

\item{emplois}{Le vecteur des emplois disponibles sur chacun des sites j (= marge des colonnes).}

\item{actifs}{Le vecteur des actifs partant de chacune des lignes visées par shuf. Le vecteur doit faire la même longueur que shuf.}

\item{f}{Le vecteur de la probabilité de fuite des actifs hors de la zone d'étude.}

\item{shuf}{Le vecteur de priorité des actifs pour choisir leur site d'arrivée. Il est possible de segmenter les départs d'une ligne i en répétant cette ligne à plusieurs endroits du shuf et en répartissant les poids au sein du vecteurs actifs.}

\item{attraction}{Choix de la fonction d'attraction des différents sites, appliquée à l'accessibilité. 
Par défaut, "constant" où aucun site n'a plus d'attrait qu'un autre. 
"marche" où l'attrait vaut béta jusqu'à une certaine distance (param alpha) puis 1. f(x) = béta si x < alpha, = 1 si x > alpha.
"logistique" où l'attrait décroît selon une fonction logistique avec une distance de bascule (param alpha), une vitesse de bascule (param béta) 
et un seuil (param gamma). Si h(x) = exp( (x-alpha)/béta), f(x) = gamma + h(x) / (1 + h(x)).
"odds" où chaque flux (from, to) se voit attribuer un odds. Dans ce cas, on entre un Row Sparse Matrix des log(odds) selon ses éléments.}

\item{param}{Dans les cas de "marche" et "logistique", un vecteur avec dans l'ordre les valeurs des paramètres alpha, béta et, si nécessaire, gamma.}

\item{modds}{un triplet au format data.frame des log(odds).}

\item{nthreads}{Nombre de threads pour OpenMP. Default : 0 = choix auto.}

\item{progress}{Ajoute une barre de progression. Default : true.}

\item{normalisation}{Calage des emplois disponibles sur le nombre d'actifs travaillant sur la zone. Default : false.}

\item{fuite_min}{Seuil minimal pour la fuite d'un actif. Doit être supérieur à 0. Défault = 1e-3.}
}
\value{
renvoie un triplet au format data.frame des estimations du nombre de trajets de i vers j.
}
\description{
La fonction meaps sur plusieurs shufs selon le mode de calcul continu.
}
