% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{meaps_opt}
\alias{meaps_opt}
\title{Refonte de l'optim en wrapper
STRATEGIE :
1 = cmaes::cma_es = Covariance Matrix Adapting Evolutionary Strategy.
2 = dfoptim::nmkb = Nelder-Mead derivative free algorithm.
3 = dfoptim::mads = mesh adaptative direct search.}
\usage{
meaps_opt(
  MeapsDataGroup,
  attraction,
  parametres,
  fct_meaps = "all_in",
  strategie = 1L,
  amplitude_max = 1000,
  rayon_max = 10,
  lower = NULL,
  upper = NULL,
  control = NULL,
  nthreads = 0L,
  progress = TRUE,
  objective = "kl",
  quiet = TRUE
)
}
\description{
Refonte de l'optim en wrapper
STRATEGIE :
1 = cmaes::cma_es = Covariance Matrix Adapting Evolutionary Strategy.
2 = dfoptim::nmkb = Nelder-Mead derivative free algorithm.
3 = dfoptim::mads = mesh adaptative direct search.
}
