% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{meaps_optim}
\alias{meaps_optim}
\title{La fonction meaps_optim sert pour la recherche des meilleurs paramètres ou odds lorsqu'un certain regroupement des flux est connu.}
\usage{
meaps_optim(
  dist_prep,
  emplois,
  actifs,
  fuite,
  shuf,
  groups_from,
  groups_to,
  attraction = "constant",
  param = 0,
  odds_prep = NULL,
  nthreads = 0,
  progress = TRUE,
  normalisation = FALSE,
  fuite_min = 0.001
)
}
\arguments{
\item{dist_prep}{Matrice des distances au format RankedRSMatrix (sortie de la fonction prep_meaps_dist).}

\item{emplois}{Le vecteur des emplois disponibles sur chacun des sites j (= marge des colonnes).}

\item{actifs}{Le vecteur des actifs partant de chacune des lignes visées par shuf. Le vecteur doit faire la même longueur que shuf.}

\item{shuf}{Le vecteur de priorité des actifs pour choisir leur site d'arrivée. Il est possible de segmenter les départs d'une ligne i en répétant cette ligne à plusieurs endroits du shuf et en répartissant les poids au sein du vecteurs actifs.}

\item{attraction}{choix de la fonction d'attraction. Default = "constant". "marche", "logistique" ou "odds".}

\item{param}{paramètres pour mode = "marche" (le 1er est la distance de la marche, le 2nd est le plancher après la marche).
pour mode = "logistique" (le 1er est la distance du point de symétrie, le 2nd la raideur de la bascule, le 3ème le plancher).}

\item{odds_prep}{les log(odds) au format RankedRSMatrix, mais rangé selon dist (résultat de prep_meaps_odds).}

\item{nthreads}{Nombre de threads pour OpenMP. Default : 0 = choix auto.}

\item{progress}{Ajoute une barre de progression. Default : true.}

\item{normalisation}{Calage des emplois disponibles sur le nombre d'actifs travaillant sur la zone. Default : false.}

\item{fuite_min}{Seuil minimal pour la fuite d'un actif. Doit être supérieur à 0. Défault = 1e-3.}

\item{f}{Le vecteur de la probabilité de fuite des actifs hors de la zone d'étude.}

\item{group_from}{Le vecteur de regroupement des communes de départs (lignes).}

\item{group_to}{Le vecteur de regroupement des communes de destinations (colonnes).}
}
\value{
renvoie une matrice avec les estimations des flux regroupés.
}
\description{
La fonction meaps_optim sert pour la recherche des meilleurs paramètres ou odds lorsqu'un certain regroupement des flux est connu.
}
