% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{meaps_optimx}
\alias{meaps_optimx}
\title{version optimx.
Beaucoup plus de méthodes, dont certaines recommandées : nvm(), ncg(), ucminf::ucminf(), lbfgs::lbfgs(), nloptr::lbfgs()}
\usage{
meaps_optimx(
  MeapsDataGroup,
  attraction,
  parametres,
  version = "all_in",
  method = "lbfgsb3c",
  objective = "KL",
  amplitude_max = 1000,
  rayon_max = 10,
  lower = NULL,
  upper = NULL,
  control = NULL,
  discret = NULL,
  nthreads = 0L,
  progress = TRUE,
  quiet = TRUE
)
}
\description{
version optimx.
Beaucoup plus de méthodes, dont certaines recommandées : nvm(), ncg(), ucminf::ucminf(), lbfgs::lbfgs(), nloptr::lbfgs()
}
