% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meaps_single}
\alias{meaps_single}
\title{La fonction meaps sur un shuf.}
\usage{
meaps_single(rkdist, emplois, actifs, modds, f, shuf)
}
\arguments{
\item{rkdist}{La matrice des rangs dans lequel les colonnes j sont passées en revue pour chacune des lignes i.}

\item{emplois}{Le vecteur des emplois disponibles sur chacun des sites j (= marge des colonnes).}

\item{actifs}{Le vecteur des actifs partant de chacune des lignes visées par shuf.}

\item{modds}{La matrice des odds modifiant la chance d'absorption de chacun des sites j pour des résidents en i.}

\item{f}{Le vecteur de la probabilité de fuite des actifs hors de la zone d'étude.}

\item{shuf}{Le vecteur de priorité des actifs pour choisir leur site d'arrivée. 
Il est possible de segmenter les départs d'une ligne i en répétant cette ligne à plusieurs endroits du shuf.
Dans ce cas, le nombre d'actifs sera répartie également entre les différents départs depuis cette ligne.}
}
\value{
renvoie une matrice avec les estimations du nombre de trajets de i vers j.
}
\description{
La fonction meaps sur un shuf.
}
