% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\name{meapsdata}
\alias{meapsdata}
\title{Constructeur de MeapsData.}
\usage{
meapsdata(
  triplet,
  actifs,
  emplois,
  fuites,
  nshuf = NULL,
  seuil = 40,
  seed = NULL
)
}
\arguments{
\item{triplet}{Un triplet (fromidINS, toidINS, metric) trié lexicographiquement selon fromidINS et metric.}

\item{actifs}{Un vecteur du nombre d'actifs, labélisé par fromidINS.}

\item{emplois}{Un vecteur du nombre d'emplois, labélisé par toidINS.}

\item{fuites}{Un vecteur de la proportion de fuite, labélisé selon fromidINS.}
}
\description{
Constructeur de MeapsData.
}
