% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\docType{class}
\name{MeapsDataGroup-class}
\alias{MeapsDataGroup-class}
\title{Définition d'une sous-classe MeapsDataGroup qui ajoute à MeapsData des données pour regrouper les zones de départs et d'arrivées (notamment par IRIS ou par commune).}
\description{
Définition d'une sous-classe MeapsDataGroup qui ajoute à MeapsData des données pour regrouper les zones de départs et d'arrivées (notamment par IRIS ou par commune).
}
\section{Slots}{

\describe{
\item{\code{group_from}}{Le vecteur des codes de regroupement des actifs (à l'iris ou à la commune) labélisé selon les fromidINS.}

\item{\code{group_to}}{Le vecteur des codes de regroupement des emplois (à l'iris ou à la commune) labélisé selon les toidINS.}

\item{\code{cible}}{Un triplet (avec les colonnes group_from, group_to et value) des flux attendus.}

\item{\code{index_gfrom}}{Vecteur des indices correspondant à group_from.}

\item{\code{index_gto}}{Vecteur des indices correpsondant à group_to.}
}}

