% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\name{meapsdatagroup}
\alias{meapsdatagroup}
\title{Constructeur de MeapsDataGroup.}
\usage{
meapsdatagroup(MeapsData, group_from, group_to, cible)
}
\arguments{
\item{MeapsData}{Un objet MeapsData qui décrit la zone d'étude.}

\item{group_from}{Un vecteur des codes de regroupement, labélisé par fromidINS.}

\item{group_to}{Un vecteur des codes de regroupement, labélisé par toidINS.}

\item{cible}{Un triplet (group_from, group_to, value) décrivant les flux groupés de référence.}
}
\description{
Constructeur de MeapsDataGroup.
}
