% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\docType{class}
\name{MeapsDataMultiChamps-class}
\alias{MeapsDataMultiChamps-class}
\title{Définition d'une sous-classe MeapsDataMultiChamps qui ajoute à MeapsData des données synthétiques de différentes sources.}
\description{
Définition d'une sous-classe MeapsDataMultiChamps qui ajoute à MeapsData des données synthétiques de différentes sources.
}
\section{Slots}{

\describe{
\item{\code{struct_groups}}{Matrice où chaque ligne indique l'appartenance de groupe de l'élément jr dans la représentation en triplet.}

\item{\code{n_groups}}{Vecteur donnant le nombre de groupes dans chacune des lignes de struct_group.}

\item{\code{cibles}}{Liste des données agrégées sur lesquelles on souhaite se caler.}
}}

