% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class.R
\name{meapsdatamultichamps}
\alias{meapsdatamultichamps}
\title{Constructeur de MeapsDataMultiChamps.}
\usage{
meapsdatamultichamps(MeapsData, df_groups, cibles)
}
\arguments{
\item{MeapsData}{Un objet MeapsData qui décrit la zone d'étude.}

\item{df_groups}{Une data.frame avec fromidINS et toidINS dans le même ordre que le triplet MeapsData, puis avec les variables de regroupement.}

\item{cibles}{Une liste de vecteurs, chacun ayant la taille du nombre de groupes et les valeurs de flux attendues dans le groupe correspondant.}
}
\description{
Constructeur de MeapsDataMultiChamps.
}
