% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class wrapppers.R
\name{multishuf}
\alias{multishuf}
\title{Multishuf, l'original
algorithme original meaps, avec des odds discrets et quand même bien présents,
et un shuf, avec openMP}
\usage{
multishuf(
  MeapsData,
  attraction = "constant",
  parametres = 0,
  odds = 1,
  nshuf = 16,
  nthreads = 0L,
  verbose = TRUE
)
}
\arguments{
\item{MeapsData}{}

\item{attraction}{}

\item{parametres}{}

\item{odds}{}

\item{nshuf}{}

\item{nthreads}{}

\item{verbose}{}
}
\description{
Multishuf, l'original
algorithme original meaps, avec des odds discrets et quand même bien présents,
et un shuf, avec openMP
}
