% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class wrapppers.R
\name{multishuf_oc}
\alias{multishuf_oc}
\title{Multishuf, la version continue
algorithme original meaps, 2eme version, avec des odds continus,
et un shuf, avec openMP}
\usage{
multishuf_oc(
  MeapsData,
  attraction = "constant",
  parametres = 0,
  odds = 1,
  nshuf = 16,
  nthreads = 0L,
  verbose = TRUE,
  gbperthreads = 4
)
}
\arguments{
\item{MeapsData}{}

\item{attraction}{}

\item{parametres}{}

\item{odds}{}

\item{nshuf}{}

\item{nthreads}{}

\item{verbose}{}
}
\description{
Multishuf, la version continue
algorithme original meaps, 2eme version, avec des odds continus,
et un shuf, avec openMP
}
