% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_class wrapppers.R
\name{multishuf_oc_grouped}
\alias{multishuf_oc_grouped}
\title{Multishuf continu groupé}
\usage{
multishuf_oc_grouped(
  MeapsDataGroup,
  attraction = "constant",
  parametres = 0,
  odds = 1,
  nthreads = 0L,
  verbose = TRUE
)
}
\arguments{
\item{MeapsDataGroup}{object avec l'information de groupe}

\item{attraction}{fonction pour calculer les odds}

\item{parametres}{paramètre de la fonction "attraction"}

\item{odds}{possibilité de passer un vecteur d'odds complet}

\item{nthreads}{nombre de threads}

\item{verbose}{affiche des messages}
}
\description{
Calcule les flux meaps à partir de l'algorithme multishuf et
retourne un résultat agrégé selon les groupes envoyés
}
