% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{multishuf_origin}
\alias{multishuf_origin}
\title{Multishuf, l'original dans sa version discrête
algorithme original meaps, avec des odds discrets et quand même bien présents,
et un shuf, avec openMP}
\usage{
multishuf_origin(
  MeapsData,
  attraction = "constant",
  parametres = 0,
  odds = 1,
  nshuf = 16,
  nthreads = 0L,
  verbose = TRUE
)
}
\arguments{
\item{MeapsData}{}

\item{verbose}{}
}
\description{
Multishuf, l'original dans sa version discrête
algorithme original meaps, avec des odds discrets et quand même bien présents,
et un shuf, avec openMP
}
