% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{multishuf_task}
\alias{multishuf_task}
\title{Multishuf, réécrit en version continue avec une clause omp task depend.}
\usage{
multishuf_task(
  MeapsData,
  attraction = "constant",
  parametres = 0,
  nshuf = 16,
  seuil = 40,
  nthreads = 0L,
  verbose = TRUE
)
}
\description{
Multishuf, réécrit en version continue avec une clause omp task depend.
}
