% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{multishuf_task_grouped}
\alias{multishuf_task_grouped}
\title{Fonction all_in groupé.
Si MeapsDataGroup a défini une cible, renvoie également l'entropie relative (kl).}
\usage{
multishuf_task_grouped(
  MeapsDataGroup,
  attraction = "constant",
  parametres = 0,
  odds = 1,
  nthreads = 0L,
  verbose = TRUE
)
}
\description{
Fonction all_in groupé.
Si MeapsDataGroup a défini une cible, renvoie également l'entropie relative (kl).
}
