% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_methods.R
\name{ponderer_champs}
\alias{ponderer_champs}
\title{Fonction de choix d'une pondération pour optimiser MeapsDataMultiChamps
On évalue la part d'information qu'apporte chaque source par rapport à une estimation Meaps à attraction constante.}
\usage{
ponderer_champs(MeapsDataMultiChamps, fct_meaps = "all_in")
}
\description{
Fonction de choix d'une pondération pour optimiser MeapsDataMultiChamps
On évalue la part d'information qu'apporte chaque source par rapport à une estimation Meaps à attraction constante.
}
