% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{prep_meaps_masses}
\alias{prep_meaps_masses}
\title{La fonction prep_meaps_odds prépare la matrice des odds pour traitement par meaps_optim.}
\usage{
prep_meaps_masses(emplois, actifs, fuite, shuf, cle_from, cle_to)
}
\arguments{
\item{cle_from}{la clé_from issue de distance.}

\item{cle_to}{la clé to}

\item{modds}{un triplet des log odds.}
}
\value{
renvoie une RankedRSMatrix des odds, rangés selon le rang des distances.
}
\description{
La fonction prep_meaps_odds prépare la matrice des odds pour traitement par meaps_optim.
}
