% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{prep_meaps_odds}
\alias{prep_meaps_odds}
\title{La fonction prep_meaps_odds prépare la matrice des odds pour traitement par meaps_optim.}
\usage{
prep_meaps_odds(modds, dist)
}
\arguments{
\item{modds}{une matrice des odds, une dgRMatrix des log(odds) ou la RankedRSMatrix rangée selon dist.}

\item{dist}{Matrice des distances où des résidents en ligne i rejoignent des opportunités en colonnes j.}
}
\value{
renvoie une RankedRSMatrix des odds, rangés selon le rang des distances.
}
\description{
La fonction prep_meaps_odds prépare la matrice des odds pour traitement par meaps_optim.
}
