% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{prep_meaps_odds_on_dist}
\alias{prep_meaps_odds_on_dist}
\title{La fonction prep_meaps_odds_on_dist prépare la matrice des odds pour traitement par meaps_optim.
c'est une matrice ranké, mais selon les distances}
\usage{
prep_meaps_odds_on_dist(odds, prep)
}
\arguments{
\item{odds}{un triplet des odds.}

\item{prep}{une liste de préparation issue de prep_meaps_dist.}
}
\value{
renvoie une RankedRSMatrix des odds, rangés selon le rang des distances.
}
\description{
La fonction prep_meaps_odds_on_dist prépare la matrice des odds pour traitement par meaps_optim.
c'est une matrice ranké, mais selon les distances
}
