% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2kl2}
\alias{r2kl2}
\title{R2 sur la base de l'entropie relative
On prend comme référence une distribution uniforme
On calcule le KL de cette distribution et de la distributiopn observée (référence)
Puis le KL de la distribution prposée et de la distribution de référence
on défini alors R2 = 1-kl(est, obs)/kl(unif, obs)
Attention ! On exclue les valeurs nulles des deux distributions}
\usage{
r2kl2(estime, observe, seuil = 0.99, bruit = 1e-06)
}
\arguments{
\item{estime}{distribution empirique estimée}

\item{observe}{distribution empirique observée}

\item{seuil}{Seuil de la distribution observée cumulée au dessus duquel on aggrège les effectifs. Défaut : 99%.}

\item{bruit}{remplace les valeurs inférieures au bruit par ce bruit. Défaut 1e-6}
}
\value{
une valeur de R2 filtré ($r2kl) et une non filtrée
}
\description{
R2 sur la base de l'entropie relative
On prend comme référence une distribution uniforme
On calcule le KL de cette distribution et de la distributiopn observée (référence)
Puis le KL de la distribution prposée et de la distribution de référence
on défini alors R2 = 1-kl(est, obs)/kl(unif, obs)
Attention ! On exclue les valeurs nulles des deux distributions
}
