% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{shannon}
\alias{shannon}
\title{Calcul de l'entropie de Shannon.}
\usage{
shannon(nb_ref, seuil_collapse = 1)
}
\arguments{
\item{nb_ref}{Vecteur dont on veut mesurer l'entropie.}

\item{seuil_collapse}{Seuil en dessous duquel on aggrège les effectifs. Défaut : 1.}
}
\value{
Entropie.
}
\description{
Calcul de l'entropie de Shannon.
}
