% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{triplet2listij}
\alias{triplet2listij}
\title{Fonction de transformation d'un triplet fromidINS toidINS value à une liste avec i j x et les clés de correspondances.}
\usage{
triplet2listij(data)
}
\description{
Fonction de transformation d'un triplet fromidINS toidINS value à une liste avec i j x et les clés de correspondances.
}
