% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tripletlodds2dgr}
\alias{tripletlodds2dgr}
\title{Fonction de transformation d'un triplet de log(odds) avec fromidINS et toidINS en une matrice row sparse.
Les cle_from et cle_to sont celles tirées du triplet des distances.}
\usage{
tripletlodds2dgr(data, cle_from, cle_to)
}
\description{
Fonction de transformation d'un triplet de log(odds) avec fromidINS et toidINS en une matrice row sparse.
Les cle_from et cle_to sont celles tirées du triplet des distances.
}
