% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsdata_checker.R
\name{is_triplet_meaps}
\alias{is_triplet_meaps}
\title{Définition d'une classe d'objet rmeaps de préparation des données.
Fonctions de validation de MeapsData et MeapsDataGroup}
\usage{
is_triplet_meaps(object, quiet = TRUE)
}
\description{
Définition d'une classe d'objet rmeaps de préparation des données.
Fonctions de validation de MeapsData et MeapsDataGroup
}
