% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{meaps_optim}
\alias{meaps_optim}
\title{La fonction meaps_optim sert pour la recherche des meilleurs paramètres ou odds lorsqu'un certain regroupement des flux est connu.}
\usage{
meaps_optim(
  prep,
  attraction = "constant",
  param = 0,
  odds_prep = NULL,
  nthreads = 0,
  progress = FALSE,
  normalisation = FALSE,
  fuite_min = 0.001
)
}
\arguments{
\item{prep}{liste, issue de prep_meaps_dist, qui contient plusieurs éléments précalculés.}

\item{attraction}{choix de la fonction d'attraction. Default = "constant". "marche", "logistique" ou "odds".}

\item{param}{paramètres pour mode = "marche" (le 1er est la distance de la marche, le 2nd est le plancher après la marche).
pour mode = "logistique" (le 1er est la distance du point de symétrie, le 2nd la raideur de la bascule, le 3ème le plancher).}

\item{odds_prep}{les log(odds) au format RankedRSMatrix, mais rangé selon dist (résultat de prep_meaps_odds).}

\item{nthreads}{Nombre de threads pour OpenMP. Default : 0 = choix auto.}

\item{progress}{Ajoute une barre de progression. Default : true.}

\item{normalisation}{Calage des emplois disponibles sur le nombre d'actifs travaillant sur la zone. Default : false.}

\item{fuite_min}{Seuil minimal pour la fuite d'un actif. Doit être supérieur à 0. Défault = 1e-3.}
}
\value{
renvoie une matrice avec les estimations des flux regroupés.
}
\description{
La fonction meaps_optim sert pour la recherche des meilleurs paramètres ou odds lorsqu'un certain regroupement des flux est connu.
}
