% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meaps_optim.R
\name{prep_meaps_dist}
\alias{prep_meaps_dist}
\title{La fonction prep_meaps_dist prépare les données de distances pour traitement par meaps_optim}
\usage{
prep_meaps_dist(dist)
}
\arguments{
\item{dist}{triplet des distances où des résidents en ligne i rejoignent des opportunités en colonnes j.}
}
\value{
renvoie une RankedRSMatrix des distances.
}
\description{
La fonction prep_meaps_dist prépare les données de distances pour traitement par meaps_optim
}
